+++ src/assets/tokens.generated.scss
@@ -143,7 +143,6 @@ $white-70: rgb(255 255 255 / 70%);
 $white-08: rgb(255 255 255 / 8%);
 $white-base: #fff;
 $white-00: rgb(255 255 255 / 0%);
-$none-value: undefined;
 $transparent-blue-18: rgb(12 65 154 / 18%);
 $transparent-blue-08: rgb(12 65 154 / 8%);
 $transparent-blue-00: rgb(12 65 154 / 0%);
