+++ src/designTokens/tokens/amelipro/apVariantConfig.ts
@@ -1,3 +1,6 @@
+export { apColorsTokens } from './apColors'
+export { apContextualTokens } from './apContextual'
+
 export const apSemanticValues: Record<string, string> = {
 	'primary-darker-2': '$cyan-darken-60',
 	'primary-darker': '$cyan-darken-40',
